+++ portfolio/src/main/java/com/google/sps/servlets/DataServlet.java
@@ -121,7 +121,7 @@ public class DataServlet extends HttpServlet {
     String mood = (String) comment.getProperty(COMMENT_MOOD);
     String nickname = (String) comment.getProperty(COMMENT_NICKNAME);
     String imageUrl = (String) comment.getProperty(COMMENT_IMAGE_URL);
-    float sentiment = (float) comment.getProperty(COMMENT_SENTIMENT);
+    double sentiment = (double) comment.getProperty(COMMENT_SENTIMENT);
 
     return Comment.builder().setId(id).setText(text).setTimestamp(timestamp)
               .setMood(mood).setNickname(nickname).setImageUrl(imageUrl)
@@ -161,7 +161,7 @@ public class DataServlet extends HttpServlet {
     String mood = (String) request.getParameter(COMMENT_MOOD);
     String nickname = HomeServlet.userNickname();
     String imageUrl = getUploadedFileUrl(request, COMMENT_IMAGE_URL).orElse("");
-    float sentiment = calculateSentiment(text);
+    double sentiment = calculateSentiment(text);
 
     Entity commentEntity = new Entity(COMMENT_ENTITY);
     commentEntity.setProperty(COMMENT_TEXT, text);
